+++ src/cpu/cpu_constants.js
@@ -126,3 +126,4 @@ define('MASKIF', 63);
 define('WAIT', 64);
 define('PUSHIR', 65);
 define('SYNC', 66);
+define('INDEX', 67);